Break polylineFunctions into separate files
