+++ src/polylineFunctions/polylineAlloc.ts
@@ -1,0 +1,5 @@
+import { IPolyline } from "../types";
+
+export function polylineAlloc(): IPolyline {
+  return [];
+}
