+++ src/polylineFunctions/polylineClose.ts
@@ -1,0 +1,22 @@
+import { IPolyline } from "../types";
+import { polylineAlloc } from "./polylineAlloc";
+
+export function polylineClose(poly: IPolyline, out = polylineAlloc()) {
+  const len = poly.length;
+  if (len === 0) {
+    out.length = 0;
+    return out;
+  }
+
+  if (out.length !== len + 2) {
+    out.length = len + 2;
+  }
+
+  for (let i = 0; i < len; i++) {
+    out[i] = poly[i];
+  }
+
+  out[len] = poly[0];
+  out[len + 1] = poly[1];
+  return length;
+}
