+++ src/polylineFunctions/polylineGetNumSegments.ts
@@ -1,0 +1,5 @@
+import { IPolyline } from "../types";
+
+export function polylineGetNumSegments(poly: IPolyline) {
+  return poly.length / 2 - 1;
+}
