+++ src/polylineFunctions/polylineIsClosed.ts
@@ -1,0 +1,12 @@
+import { EPSILON_SQ } from "../internal/parameters";
+import { IPolyline } from "../types";
+
+export function polylineIsClosed(poly: IPolyline) {
+  if (poly.length === 0) {
+    return true;
+  } else {
+    const dx = poly[poly.length - 2] - poly[0];
+    const dy = poly[poly.length - 1] - poly[1];
+    return dx * dx + dy * dy < EPSILON_SQ;
+  }
+}
